% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/22_mcmc_sar.R
\name{sar_gibbs_sampler}
\alias{sar_gibbs_sampler}
\title{Full SAR MCMC (joint over rho & alpha, + optional covariates/factors)}
\usage{
sar_gibbs_sampler(
  Y0_pre,
  Yc_pre,
  Xc_pre,
  W,
  w,
  M = 2000,
  burn = 1000,
  verbose = TRUE,
  p_factors = 0,
  step_rho = 0.01,
  step_alpha = 0.01,
  a0 = 1,
  b0 = 1
)
}
\description{
Full SAR MCMC (joint over rho & alpha, + optional covariates/factors)
}
\keyword{internal}
